Rename & rationalize python_bindings C++ filenames
- Overlap between names of the bindings and the Halide source files was confusing
- 'python' folder contained no actual python source
- Image.h actually handles Buffer
