+++ python_bindings/src/PyHalide.cpp
@@ -1,20 +1,20 @@
 #include <boost/python.hpp>
 
-#include "Argument.h"
-#include "BoundaryConditions.h"
-#include "Error.h"
-#include "Expr.h"
-#include "Func.h"
-#include "Function.h"
-#include "IROperator.h"
-#include "Image.h"
-#include "InlineReductions.h"
-#include "Lambda.h"
-#include "Param.h"
-#include "RDom.h"
-#include "Target.h"
-#include "Type.h"
-#include "Var.h"
+#include "PyArgument.h"
+#include "PyBoundaryConditions.h"
+#include "PyError.h"
+#include "PyExpr.h"
+#include "PyFunc.h"
+#include "PyFunction.h"
+#include "PyIROperator.h"
+#include "PyBuffer.h"
+#include "PyInlineReductions.h"
+#include "PyLambda.h"
+#include "PyParam.h"
+#include "PyRDom.h"
+#include "PyTarget.h"
+#include "PyType.h"
+#include "PyVar.h"
 
 BOOST_PYTHON_MODULE(halide) {
     define_argument();
