+++ python_bindings/src/PyOperators.cpp
@@ -1,3 +1,3 @@
-#include "add_operators.h"
+#include "PyOperators.h"
 
 // nothing to define here, only templates in the headers
