+++ python_bindings/src/PyRDom.cpp
@@ -1,11 +1,11 @@
-#include "RDom.h"
+#include "PyRDom.h"
 
 #include <boost/python.hpp>
 #include <string>
 
 #include "Halide.h"
 
-#include "add_operators.h"
+#include "PyOperators.h"
 
 namespace h = Halide;
 namespace p = boost::python;
